Adding basic arithemtic operations to menu, functions working

+++ C/modularization/inc/math/module.h
@@ -1,4 +1,4 @@
 #ifndef MODULE_H_HEADER
 #define MODULE_H_HEADER
-float moduleMath(float a, float b);
+int moduleMath(int a, int b);
 #endif // MODULE_H_HEADER
